Skeletal test harness
To run, `./test.py`

The test harness sets an environment variable before importing config.py, which
tells config to use a "test environment" (store dir, key dir, and journalist
key) instead of the production environment.

I created a journalist keypair and included it in the repo for testing.
Generating a GPG keypair can be very time-consuming, and we want to be able to
run tests as fast as possible. The keypair is clearly marked so it is not used
by anyone for anything besides testing this application.