temporarily remove dangerous characters from wordlist
Until we can audit the python-gnupg library for shell injection vectors
(avoiding these was a key reason for isis lovecruft's fork), we are removing
unsafe characters from the wordlist and safelist so they are never passed to
the underlying gpg process.

This very slightly reduces the length of the wordlist which in turn reduces
codename entropy, but this tradeoff is worthwhile until further investigation
can be done.