Teach Vagrant to skip "production" Ansible tags
Here we mark various tasks with the following tags:
- production
- staging
- development

The point of this is to mark certain tasks that are specific to a particular work flow.
Therefore this "development work flow" means that the Vagrantfile specifies that it
wishes to skip tasks tagged with "production" tag.

We can of course create the inversion of this scenario where we run *only* the tags
we specify...