Update ansible's copy of config.py
Originally we had this symlinked to `securedrop/config.py.example`, but
that caused issues on Travis. For now, I'm manually updating the copy.
In the future, we should see if we can get Travis working with the
symlink so we can avoid having to remember to manually sync changes to
the application config.