building the ossec deb packages take awhile and it seems the ssh connection is timing out.
Since the ossec deb packge does not change often and we are working on migrating to building with pbuilder commenting out that role in both staging playbooks.

The app/mon roles already are configured to install ossec from the fpf repo if it wasn't already installed so this will not break anything.