Removed the the vagrant-hostmanger plugin
This is not needed for virtualbox environment since we hardcode the IP address in the Vagrantfile for each VM.

The hostmanger plugin is required when using digital ocean droplets instead of virtualbox.

Digital ocean droplets are used in our snap-ci environment.

When using digital ocean droplets the IP address of the VPS can change when you destroy and rebuild the vps. The hostmanger plugin was used with the ansible skip-tag `static-hosts` to configure the `/etc/hosts` entries