drop invalid state packets
Drop but do not log packets with invalid state. This situation was occuring around sending smtp alerts to google smtp servers. Where the connection would be removed from the iptables state table, but the mon server would then still try to send a FIN, RST packet but since it was no longer in the state table it lost the associated PID and UID. The distant end already closed the connection so this is not prevent the alerts from being sent but does cause a OSSEC spam loop. An OSSEC alert is sent causing a fw drop, that then generates another ossec alert.

These packets are droped but not logged by jumping (-j DROP) directly to DROP chain instead of the LOGNDROP chain