Don't log grsec denying rwxmap
This feature creates quite a lot of log noise. While it is worth
investigating more in depth, it does not appear to cause any breakage.
There are measures in grsec to deny these allocations and return a
useful error code to the application, and no breakage has been observed.
These messages are useful for auditing applications that make
potentially risky memory allocations, but are not useful in the
day-to-day operation of a production instance and create a lot of
logging noise. These messages are surpressed via the kernel flag until
we can investigate them further and see if there are any clear cases in
which we would want to have this logging.

Docs: https://en.wikibooks.org/wiki/Grsecurity/Appendix/Grsecurity_and_PaX_Configuration_Options#Denied_RWX_mmap.2Fmprotect_logging