add a pytest-pep8 configuration that currently does not produce pep8 failures
the max line length value is rather high, it should be set to a lower value after agreeing on some.

I personally think 80 is way to low (and outdated, we don't have vt80 terminals any more these days),
but 160 is to high. I suggest to use 120 or 132.

E402 is a matter of taste somehow, can be fixed or not IMHO.

I didn't fix the migration scripts and install files yet, thus they are currently ignored so that they do not produce test failures.