Script to migrate 0.3pre -> 0.3 (tagged)
See the README.md for more information. This script is intended to be
used to migrate 0.3pre instances to the tagged version 0.3.

This script may not yet be complete - there are still some issues with
changes made by the Ansible playbooks between installs. I am not sure if
this script is the best place to address those changes, or if we can use
Ansible to do that.