Marked logo.png as conffile
Renamed logo.png.example to logo.png and marked the file as a
configuration file in the securedrop-app-code conffiles.

Added the dpkg options to the cron-apt upgrade command in
`/etc/cron-apt/action.d/5-security` to force keeping the old conffile if
it was locally modifed while updating all other files that were not
locally modified.

For orgs that use a branded custom header image this will ensure that it
doesn't get overwritten during security updates.

Tested using a test apt repo with a 0.3.1 versioned securedrop-app-code
that cron-apt update to.