fix pre0.3 to 0.3.1 upgrade script
The pre0.3 SD version had an issue where deleting large submission would timeout, leaving the encrypted submission in place but removing it from the db. Added @garrett's fix for cleaning that up prior to running the db migration.

Re-ordered the sequence of events so the backup task and the migration happen prior to removing the old deb packages. This ensures that a custom logo.png will get backuped before dpkg deletes it.