Name migration script directory for the version it migrates *from*
When you do a migration, you're starting with an existing system with a
known version; therefore, it makes more sense to group migration scripts
by the version they are migrating *from* rather than the version they
are migration *to*. This is also better because it is conceivable for
multiple base versions (e.g. 0.2.1 *and* 0.3pre) to need migration
support to a single version (e.g. 0.3).

This creates a consistent naming scheme for all of the directories in
`migration_scripts`, because 0.3pre was already the set of scripts used
to migration 0.3pre installations to 0.3.