Fix migration script directories after merge
In `check-migration-scripts-for-0.3`, I had made the change to standardize
the directory names in `migration_scripts` to correspond to the version
one is migrating *from*. This change was not made in
`upgrade-helper-script`, so this commit makes all of the naming
consistent.

This commit deletes the old README.md because it has been subsumed by
`0.3-pre-upgrade-to-0.3.3.md`.