Make 0.3pre upgrade script idempotent for Ansible
Sometimes when the Ansible upgrade role is run, the upgrade script
completes successfully, but then something goes wrong later in the
role. For example, network connectivity issues may cause one of the
later package upgrades to fail. In this situation, we need to be able to
re-run the playbook. This means the `0.3pre_upgrade.py` script needs to
be idempotent.

The script performs a full SecureDrop backup to `/tmp` before continuing
on both the app and monitor servers. Since it would be complex to verify
whether the backups were idempotent, we simply make a backup every time
the script is run. The backup filenames are timestamped, so there is no
concern about accidentally overwriting them. Since the files are stored
in `/tmp`, they will automatically be cleaned up after the next reboot.

On the app server, the upgrade script also migrates the app database. To
make this idempotent, we check for the last change made as part of the
database migration. If that change is present, we take it as evidence
that the database has already been migrated and skip the migration. This
makes the database migration idempotent.

This commit also does some formatting cleanup, adds docstrings, and
encapsulates getting the database connection into a function to make the
script more DRY.