removes requirement for root privileges to run script
Previously, the script forced invocation as root (UID 0).
Elevated privileges are only necessary for two actions:

  apt-get install git
  apt-get install ansible

The vast majority of tasks in this script can and should be
run as the default user "amnesia", so running the script
as root complicated the logic by requiring the normal user
environment to be recreated, e.g. in SSH connections.