removes "set -u" option
This option won't work in the migration script in its present form
due to the vars '$2' and '$monitor_ip', both of which are defined
via YAML parsing, and therefore not explicitly declared within
this script. The goal of adding "set -u" was to catch
undefined (or typo'd) variables prior to running commands with
them interpolated, but the script ran successfully during testing
just now.

At a later date, we can refactor this bash script to be more
resilient if need be.