This will update the App servers torrc config file for changing the document interface tor port from 8080 to 80. As of this release the prod playbook restarts tor if the torrc config changes. This will break ansible playbook runs that are over tor. Issue #940 will track modifying the prod playbook to ensure that this isn't an issue going forward.
This does not change the ATHS onion address or secret value only the port
for the document interface will change.

This will only replace/add the line when the old HiddenServicePort config
for port 8080 exists. This will not modify the torrc on the monitor server
which only has a HiddenServicePort 22 config.