modifies vagrant_ssh_cmd to filter conn closed messages
Every `vagrant ssh -c $host` command sends a message to
STDERR when finished: "Connection to $ip_address closed."
That output is not useful, and clutters up the spectests
output, so filter for it and remove it.

Current implementation greedily silences all stderr,
but only on these ssh commands for dynamic var reading.