Re-ordered iptables ssh throttling rules
* Moved the ssh rate throttling rule above the more generic tor rule.
* Added a drop rule after the rate limiting rule for ssh
* Adds notes the ansible.cfg about relationship to the iptable rules

Other Notes:

* Directly ssh'ing into the servers only matches the allow direct access exemption rule's user ID and interface. It will not match the rate limiting rule.
* SSH'ing over tor will match the ssh rate limiting.
* Not all input rules are currently being used. The generic tor input rule will match traffic for the apt input rule before hitting. More access is not allowed, but we could organize this better.

For the ssh connections over tor after the first 3 packets with the a NEW flag are accepted by iptables the rate limiting of 3/min starts. Any packets that are above that threshold will not match the rate limiting rule and will be processed by the next rule in the list.

The next rule is a drop rule specifically for tcp packets with the NEW flag from the tor user to the local loopback.

All the traffic from the tor user to destination port 22 on the local loobpack above the rate limit threshold will match this drop rule.

This does not affect the staging environments unless testing over tor.
This does not affect the development environment at all.