Silences vagrant deb package download in test script.
Downloading and installing vagrant has been reliable,
so we don't need to the verbosity at this point. The
snap-ci web interface prints a new line for every
refresh on the download rate (likely because it doesn't
understand carriage returns), which means the first
several pages of output for the development host
are simply downloading vagrant. In the event of a failed
build, that's a lot of scrolling to get to the relevant
information.

Since we're using the `-q` flag on wget , errors with the
vagrant deb package download will still be visible.