Adjusts directory permissions check for development host.
Includes verbose comment with TODO. The problem is
the Ansible config not specifying permissions correctly,
and further complicated by the /vagrant directory
being mounted with different permissions. Likely fixing
the Ansible task that sets permissions on the logo will
resolve the permissions disparity, but will need to test
to make sure. For now, the test is passing and will fail
when the Ansible task changes, which is intended.