Update expired signing key
When we created the "Freedom of the Press Foundation Master Signing
Key", we set a 1 year expiration on the primary key, as is a common best
practice for GPG keys. We accidentally allowed it to expire on Oct 26
2015. We've updated the expiration date and published the updated public
key to the keyservers, but since the Ansible roles currently use the
copy of the public key in the Git repo (instead of fetching it from the
keyservers) during installation, we need to do a minor release to update
the copy of the public key in the repo.