Consolidate SecureDrop bashrc additions
Previously the bashrc additions that force use of a terminal multiplexer
for interactive SSH sessions were added to each admin account in
~/.bashrc. Since we want ensure the additions are sourced for all
accounts, it's cleaner to declare them in /etc/profile.d.

For backwards compatibility, added a cleanup task to remove the source
line from each admin's account. We can remove that cleanup task at a
later date, tenatively post-0.4.

Also includes corresponding spectest updates to test for the additions
in the new location, and to validate that the cleanup task worked.