Moves kernel module blacklisting into common role
Conceptually it makes sense to lump the blacklisting of kernel modules
with the other kernel-related tasks, specifically grsecurity, but in
testing environments we sometimes skip the grsecurity role, and we don't
want to skip the blacklisting in those cases. Therefore we'll include
the blacklisting as part of "common", where it will run all the time
every time.