Cleans up duplicate /etc/hosts entries
A previous version of the lineinfile task managing /etc/hosts was
needlessly appending duplicate lines. The task has been adjusted to use
a smarter regex, but that's not enough, since the new regex will update
the last match in the file, and /etc/hosts parsing returns the first
match—will could be outdated.

Therefore we'll need to duplicate the lines in the file, rewrite the
file, then perform the substitution. This more careful approach requires
three tasks rather than one, but ensures idempotence.