Fixes fetching Source Interface THS URL
Recent changes to how the Tor hidden service files are collected from
the servers and written to the Admin Workstation yielded a regression in
formatting: the "HidServAuth" string was inappropriately prepended to
all hidden service files. Instead, it should be prepended to the ATHS
files, but not to the THS file (the Source Interface).